+++ src/parser/enrichers/item/WrapsOfUnarmedPower.ts
@@ -1,3 +1,4 @@
+import { utils } from "../../../lib/_module";
 import DDBEnricherData from "../data/DDBEnricherData";
 
 export default class WrapsOfUnarmedPower extends DDBEnricherData {
@@ -9,6 +10,7 @@ export default class WrapsOfUnarmedPower extends DDBEnricherData {
   get activity(): IDDBActivityData {
     return {
       data: {
+        _id: utils.namedIDStub("wrapsOfUnarmedPower", { postfix: "core" }),
         restrictions: {
           type: "weapon",
           restriction: "unarmed",
@@ -18,7 +20,7 @@ export default class WrapsOfUnarmedPower extends DDBEnricherData {
     };
   }
 
-  bonus() {
+  get bonus() {
     const nameRegex = /Wraps of Unarmed Power, \+(\d)/;
     const match = this.name.match(nameRegex);
     if (match) {
@@ -28,7 +30,7 @@ export default class WrapsOfUnarmedPower extends DDBEnricherData {
   }
 
   get effects(): IDDBEffectHint[] {
-    const bonus = this.bonus();
+    const bonus = this.bonus;
     return [
       {
         type: "enchant",
@@ -45,8 +47,35 @@ export default class WrapsOfUnarmedPower extends DDBEnricherData {
         options: {
           description: `This unarmed attack has become a +${bonus} magic weapon, granting a bonus to attack and damage rolls.`,
         },
+        data: {
+          _id: utils.namedIDStub(`wrapsOUPow${bonus}`, { prefix: "enchant", postfix: "ef" }),
+        },
       },
     ];
+  }
+
+  get override(): IDDBOverrideData {
+
+    const flags: IDDBImporterFlags = this.ddbParser.isMuncher
+      ? {}
+      : {
+        transferEnchantment: {
+          targetItemMatches: [
+            { field: "type", value: "weapon" },
+            { field: "system.type.value", value: "natural" },
+          ],
+          effectId: utils.namedIDStub(`wrapsOUPow${this.bonus}`, { prefix: "enchant", postfix: "ef" }),
+          activityId: utils.namedIDStub("wrapsOfUnarmedPower", { postfix: "core" }),
+        },
+      };
+
+    return {
+      data: {
+        flags: {
+          ddbimporter: flags,
+        },
+      },
+    };
   }
 
 }